Fold PythonTestBuilder into the PytestRun task.
PythonTestBuilder was a vestige of the old, pre-task Pants. The PytestRun
task was implemented as a thin wrapper around it.  This change merges
the functionality into PytestRun and gets rid of PythonTestBuilder as
a standalone class.

- Fixes various kludges, such as logging:  Now code that was formerly
  in PythonTestBuilder has access to the task's logging facilities, so
  it can use those instead of print().
- PythonTestBuilder's test is now a task test for PytestRun. It uses
  the new TaskTestBase, and demonstrates how this base class could
  be used instead of the old PythonTaskTest in the remaining tests
  that still use the latter.
