+++ src/python/pants/backend/python/BUILD
@@ -28,7 +28,6 @@ python_library(
     ':python_setup',
     ':resolver',
     ':sdist_builder',
-    ':test_builder',
     ':thrift_builder',
   ]
 )
@@ -136,26 +135,6 @@ python_library(
   dependencies = [
     '3rdparty/python:pex',
   ]
-)
-
-python_library(
-  name = 'test_builder',
-  sources = ['test_builder.py'],
-  dependencies = [
-    ':python_chroot',
-    # NOTE(pl): We must pull in coverage by hand even though it isn't imported.
-    # pytest-cov otherwise pulls in a newer version of coverage that doesn't
-    # work in all environments, notably some OSX installs.
-    '3rdparty/python:coverage',
-    '3rdparty/python:pex',
-    '3rdparty/python:pytest',
-    '3rdparty/python:pytest-cov',
-    '3rdparty/python:six',
-    'src/python/pants/backend/python/targets:python',
-    'src/python/pants/base:target',
-    'src/python/pants/util:contextutil',
-    'src/python/pants/util:dirutil',
-  ],
 )
 
 python_library(
