+++ src/python/pants/backend/python/tasks/BUILD
@@ -3,8 +3,12 @@ python_library(
   sources = globs('*.py'),
   resources = globs('templates/python_eval/*.mustache'),
   dependencies = [
+    '3rdparty/python:coverage',
     '3rdparty/python:pex',
+    '3rdparty/python:pytest',
+    '3rdparty/python:pytest-cov',
     '3rdparty/python:setuptools',
+    '3rdparty/python:six',
     '3rdparty/python/twitter/commons:twitter.common.collections',
     '3rdparty/python/twitter/commons:twitter.common.dirutil',
     'src/python/pants/backend/python:antlr_builder',
@@ -12,7 +16,6 @@ python_library(
     'src/python/pants/backend/python:python_chroot',
     'src/python/pants/backend/python:python_requirement',
     'src/python/pants/backend/python:python_setup',
-    'src/python/pants/backend/python:test_builder',
     'src/python/pants/backend/python:thrift_builder',
     'src/python/pants/backend/python/targets:python',
     'src/python/pants/backend/codegen/targets:python',
