+++ tests/python/pants_test/backend/python/BUILD
@@ -4,24 +4,8 @@
 target(
   name='python',
   dependencies=[
-    ':test_builder',
     ':test_python_requirement_list',
     'tests/python/pants_test/backend/python/tasks'
-  ]
-)
-
-python_tests(
-  name='test_builder',
-  sources=['test_test_builder.py'],
-  dependencies=[
-    '3rdparty/python:coverage',
-    '3rdparty/python:pex',
-    'src/python/pants/backend/python/targets:python',
-    'src/python/pants/backend/python:python_setup',
-    'src/python/pants/backend/python:test_builder',
-    'src/python/pants/base:build_file_aliases',
-    'src/python/pants/util:contextutil',
-    'tests/python/pants_test:base_test'
   ]
 )
