+++ tests/python/pants_test/backend/python/tasks/test_pytest_run.py
@@ -11,84 +11,80 @@ import xml.dom.minidom as DOM
 from textwrap import dedent
 
 import coverage
-from pex.interpreter import PythonInterpreter
 
-from pants.backend.python.interpreter_cache import PythonInterpreterCache
-from pants.backend.python.python_setup import PythonRepos, PythonSetup
 from pants.backend.python.targets.python_library import PythonLibrary
 from pants.backend.python.targets.python_tests import PythonTests
-from pants.backend.python.test_builder import PythonTestBuilder
+from pants.backend.python.tasks.pytest_run import PytestRun, PythonTestFailure
 from pants.base.build_file_aliases import BuildFileAliases
 from pants.util.contextutil import environment_as, pushd
-from pants_test.base_test import BaseTest
+from pants_test.task_test_base import TaskTestBase
 
 
-class PythonTestBuilderTestBase(BaseTest):
+class PythonTestBuilderTestBase(TaskTestBase):
+  @classmethod
+  def task_type(cls):
+    return PytestRun
+
   def setUp(self):
     super(PythonTestBuilderTestBase, self).setUp()
     self.set_options_for_scope('', python_chroot_requirements_ttl=1000000000)
 
-  def _cache_current_interpreter(self):
-    cache = PythonInterpreterCache(PythonSetup(self.config()), PythonRepos(self.config()))
-
-    # We only need to cache the current interpreter, avoid caching for every interpreter on the
-    # PATH.
-    current_interpreter = PythonInterpreter.get()
-    for cached_interpreter in cache.setup(paths=[current_interpreter.binary]):
-      if cached_interpreter == current_interpreter:
-        return cached_interpreter
-    raise RuntimeError('Could not find suitable interpreter to run tests.')
-
-  def run_tests(self, targets, args=None, fast=True, debug=False):
-    test_builder = PythonTestBuilder(
-        self.context(),
-        targets, args or [], fast=fast, debug=debug, interpreter=self._cache_current_interpreter())
-
+  def run_tests(self, targets, debug=False):
+    options = {
+      # TODO: Clean up this hard-coded interpreter constraint once we have subsystems
+      # and can simplify InterpreterCache and PythonSetup.
+      'interpreter': ['CPython>=2.7,<3'],  # These tests don't pass on Python 3 yet.
+      'colors': False,
+      'level': 'debug' if debug else 'level'
+    }
+    self.set_options(**options)
+    context = self.context(target_roots=targets)
+    pytest_run_task = self.create_task(context)
     with pushd(self.build_root):
-      return test_builder.run()
+      pytest_run_task.execute()
 
+  def run_failing_tests(self, targets, debug=False):
+    with self.assertRaises(PythonTestFailure):
+      self.run_tests(targets=targets, debug=debug)
 
 class PythonTestBuilderTestEmpty(PythonTestBuilderTestBase):
   def test_empty(self):
-    self.assertEqual(0, self.run_tests(targets=[]))
+    self.run_tests(targets=[])
 
 
 class PythonTestBuilderTest(PythonTestBuilderTestBase):
   @property
   def alias_groups(self):
-    return BuildFileAliases.create(
-        targets={
-            'python_library': PythonLibrary,
-            'python_tests': PythonTests
-        })
+    return BuildFileAliases.create(targets={
+      'python_tests': PythonTests, 'python_library': PythonLibrary})
 
   def setUp(self):
     super(PythonTestBuilderTest, self).setUp()
 
     self.create_file(
         'lib/core.py',
-        dedent('''
+        dedent("""
           def one():  # line 1
             return 1  # line 2
                       # line 3
                       # line 4
           def two():  # line 5
             return 2  # line 6
-        ''').strip())
+        """).strip())
     self.add_to_build_file(
         'lib',
-        dedent('''
+        dedent("""
           python_library(
             name='core',
             sources=[
               'core.py'
             ]
           )
-        '''))
+        """))
 
     self.create_file(
         'tests/test_core_green.py',
-        dedent('''
+        dedent("""
           import unittest2 as unittest
 
           import core
@@ -96,18 +92,18 @@ class PythonTestBuilderTest(PythonTestBuilderTestBase):
           class CoreGreenTest(unittest.TestCase):
             def test_one(self):
               self.assertEqual(1, core.one())
-        '''))
+        """))
     self.create_file(
         'tests/test_core_red.py',
-        dedent('''
+        dedent("""
           import core
 
           def test_two():
             assert 1 == core.two()
-        '''))
+        """))
     self.add_to_build_file(
         'tests',
-        dedent('''
+        dedent("""
           python_tests(
             name='green',
             sources=[
@@ -158,20 +154,20 @@ class PythonTestBuilderTest(PythonTestBuilderTestBase):
               'core'
             ]
           )
-        '''))
+        """))
     self.green = self.target('tests:green')
     self.red = self.target('tests:red')
     self.all = self.target('tests:all')
     self.all_with_coverage = self.target('tests:all-with-coverage')
 
   def test_green(self):
-    self.assertEqual(0, self.run_tests(targets=[self.green]))
+    self.run_tests(targets=[self.green])
 
   def test_red(self):
-    self.assertEqual(1, self.run_tests(targets=[self.red]))
+    self.run_failing_tests(targets=[self.red])
 
   def test_mixed(self):
-    self.assertEqual(1, self.run_tests(targets=[self.green, self.red]))
+    self.run_failing_tests(targets=[self.green, self.red])
 
   def test_junit_xml(self):
     # We expect xml of the following form:
@@ -184,7 +180,7 @@ class PythonTestBuilderTest(PythonTestBuilderTestBase):
 
     report_basedir = os.path.join(self.build_root, 'dist', 'junit')
     with environment_as(JUNIT_XML_BASE=report_basedir):
-      self.assertEqual(1, self.run_tests(targets=[self.red, self.green]))
+      self.run_failing_tests(targets=[self.red, self.green])
 
       files = glob.glob(os.path.join(report_basedir, '*.xml'))
       self.assertEqual(1, len(files))
@@ -219,29 +215,29 @@ class PythonTestBuilderTest(PythonTestBuilderTestBase):
     self.assertFalse(os.path.isfile(self.coverage_data_file()))
     covered_file = os.path.join(self.build_root, 'lib', 'core.py')
     with environment_as(PANTS_PY_COVERAGE='1'):
-      self.assertEqual(0, self.run_tests(targets=[self.green]))
+      self.run_tests(targets=[self.green])
       all_statements, not_run_statements = self.load_coverage_data(covered_file)
       self.assertEqual([1, 2, 5, 6], all_statements)
       self.assertEqual([6], not_run_statements)
 
-      self.assertEqual(1, self.run_tests(targets=[self.red]))
+      self.run_failing_tests(targets=[self.red])
       all_statements, not_run_statements = self.load_coverage_data(covered_file)
       self.assertEqual([1, 2, 5, 6], all_statements)
       self.assertEqual([2], not_run_statements)
 
-      self.assertEqual(1, self.run_tests(targets=[self.green, self.red]))
+      self.run_failing_tests(targets=[self.green, self.red])
       all_statements, not_run_statements = self.load_coverage_data(covered_file)
       self.assertEqual([1, 2, 5, 6], all_statements)
       self.assertEqual([], not_run_statements)
 
       # The all target has no coverage attribute and the code under test does not follow the
       # auto-discover pattern so we should get no coverage.
-      self.assertEqual(1, self.run_tests(targets=[self.all]))
+      self.run_failing_tests(targets=[self.all])
       all_statements, not_run_statements = self.load_coverage_data(covered_file)
       self.assertEqual([1, 2, 5, 6], all_statements)
       self.assertEqual([1, 2, 5, 6], not_run_statements)
 
-      self.assertEqual(1, self.run_tests(targets=[self.all_with_coverage]))
+      self.run_failing_tests(targets=[self.all_with_coverage])
       all_statements, not_run_statements = self.load_coverage_data(covered_file)
       self.assertEqual([1, 2, 5, 6], all_statements)
       self.assertEqual([], not_run_statements)
@@ -251,13 +247,13 @@ class PythonTestBuilderTest(PythonTestBuilderTestBase):
     covered_file = os.path.join(self.build_root, 'lib', 'core.py')
     with environment_as(PANTS_PY_COVERAGE='modules:does_not_exist,nor_does_this'):
       # modules: should trump .coverage
-      self.assertEqual(1, self.run_tests(targets=[self.green, self.red]))
+      self.run_failing_tests(targets=[self.green, self.red])
       all_statements, not_run_statements = self.load_coverage_data(covered_file)
       self.assertEqual([1, 2, 5, 6], all_statements)
       self.assertEqual([1, 2, 5, 6], not_run_statements)
 
     with environment_as(PANTS_PY_COVERAGE='modules:core'):
-      self.assertEqual(1, self.run_tests(targets=[self.all]))
+      self.run_failing_tests(targets=[self.all])
       all_statements, not_run_statements = self.load_coverage_data(covered_file)
       self.assertEqual([1, 2, 5, 6], all_statements)
       self.assertEqual([], not_run_statements)
@@ -267,13 +263,13 @@ class PythonTestBuilderTest(PythonTestBuilderTestBase):
     covered_file = os.path.join(self.build_root, 'lib', 'core.py')
     with environment_as(PANTS_PY_COVERAGE='paths:does_not_exist/,nor_does_this/'):
       # paths: should trump .coverage
-      self.assertEqual(1, self.run_tests(targets=[self.green, self.red]))
+      self.run_failing_tests(targets=[self.green, self.red])
       all_statements, not_run_statements = self.load_coverage_data(covered_file)
       self.assertEqual([1, 2, 5, 6], all_statements)
       self.assertEqual([1, 2, 5, 6], not_run_statements)
 
     with environment_as(PANTS_PY_COVERAGE='paths:core.py'):
-      self.assertEqual(1, self.run_tests(targets=[self.all], debug=True))
+      self.run_failing_tests(targets=[self.all], debug=True)
       all_statements, not_run_statements = self.load_coverage_data(covered_file)
       self.assertEqual([1, 2, 5, 6], all_statements)
       self.assertEqual([], not_run_statements)
